backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.9.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
docker==7.1.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
psycopg2-binary==2.9.11
pillow==11.3.0
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.3
requests==2.32.4
rpds-py==0.27.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
